refactor: restructuring code
extracting functions from events, moving packages
